:card_file_box: work on relationship between tables

+++ back-end/controllers/article-ctrl.js
@@ -2,6 +2,8 @@ const db = require("../models")
 const Article = db.articles
 // const Op = db.Sequelize.Op
 const fs = require("fs")
+const jwt = require("jsonwebtoken")
+const User = db.users
 
 
 // Create a new Article
@@ -41,8 +43,21 @@ exports.getAll = (req, res) => {
 			res.send(data)
 		})
 		.catch((error) => res.status(403).json({error}))
-
 }
+
+// test pour relier les tables users et articles
+// exports.getAll = (req, res) => {
+// 	Article.findAll({
+// 		include: [{
+// 			model: User,
+// 			where: {id: 36 }
+// 			}]
+// 		})
+// 			.then((data) => {
+// 				res.send(data)
+// 			})
+// 			.catch((error) => res.status(403).json("das une errur"))
+// }
 
 // Get one article
 exports.getOne = (req, res) => {
@@ -53,17 +68,31 @@ exports.getOne = (req, res) => {
 
 // Delete one article 
 exports.delete = (req, res) => {
-	Article.findOne({id: req.params.id})
+
+	// check user id and author id
+	const usertoken = req.headers.authorization;
+	const token = usertoken.split(' ');
+	const decodedId = jwt.verify(token[1], "monTokenSuperSecret1984");
+	console.log(decodedId);
+
+	Article.findOne({id: req.params.id} && {id:decodedId})
 		.then ((article) => { 
+
 			console.log("article trouvé")  // a supprimer
-			const filename = article.picture.split("/images/")[1] 
-			console.log(filename)
-			// delete picture then delete article
-			fs.unlink(`./uploads/${filename}`, () => {
-				Article.destroy({where : {id: req.params.id}})
-					.then(() => res.status(200).json("Article supprimé")) 
-					.catch((error) => res.status(403).json({error})) 
-			})
+
+			if (article.author === decodedId) {
+				const filename = article.picture.split("/images/")[1] 
+				console.log(filename)
+				// delete picture then delete article
+				fs.unlink(`./uploads/${filename}`, () => {
+					Article.destroy({where : {id: req.params.id}})
+						.then(() => res.status(200).json("Article supprimé")) 
+						.catch((error) => res.status(403).json({error})) 
+				})
+			}
+
+			console.log("post finden but authentificaton error")  // a supprimer
+			
 		})
 		.catch((error) => res.status(403).json({error})) 
 }
